US3508: build out getRooms, getRoom, updateRoom
